+++ apps/kyb-app/CHANGELOG.md
@@ -4,10 +4,9 @@
 
 ### Patch Changes
 
-- version bump
 - Updated dependencies
-  - @ballerine/workflow-browser-sdk@0.6.116
   - @ballerine/common@0.9.93
+  - @ballerine/workflow-browser-sdk@0.6.116
 
 ## 0.3.167